+++ src/components/select/select.styles.ts
@@ -14,7 +14,7 @@ export default css`
     display: block;
   }
 
-  .select__box {
+  .select__control {
     display: inline-flex;
     align-items: center;
     justify-content: start;
@@ -31,13 +31,13 @@ export default css`
     cursor: pointer;
   }
 
-  .select:not(.select--disabled) .select__box:hover {
+  .select:not(.select--disabled) .select__control:hover {
     background-color: rgb(var(--sl-input-background-color-hover));
     border-color: rgb(var(--sl-input-border-color-hover));
     color: rgb(var(--sl-input-color-hover));
   }
 
-  .select.select--focused:not(.select--disabled) .select__box {
+  .select.select--focused:not(.select--disabled) .select__control {
     background-color: rgb(var(--sl-input-background-color-focus));
     border-color: rgb(var(--sl-input-border-color-focus));
     box-shadow: var(--sl-focus-ring);
@@ -45,7 +45,7 @@ export default css`
     color: rgb(var(--sl-input-color-focus));
   }
 
-  .select--disabled .select__box {
+  .select--disabled .select__control {
     background-color: rgb(var(--sl-input-background-color-disabled));
     border-color: rgb(var(--sl-input-border-color-disabled));
     color: rgb(var(--sl-input-color-disabled));
@@ -141,7 +141,7 @@ export default css`
    */
 
   /* Small */
-  .select--small .select__box {
+  .select--small .select__control {
     border-radius: var(--sl-input-border-radius-small);
     font-size: var(--sl-input-font-size-small);
     min-height: var(--sl-input-height-small);
@@ -184,7 +184,7 @@ export default css`
   }
 
   /* Medium */
-  .select--medium .select__box {
+  .select--medium .select__control {
     border-radius: var(--sl-input-border-radius-medium);
     font-size: var(--sl-input-font-size-medium);
     min-height: var(--sl-input-height-medium);
@@ -227,7 +227,7 @@ export default css`
   }
 
   /* Large */
-  .select--large .select__box {
+  .select--large .select__control {
     border-radius: var(--sl-input-border-radius-large);
     font-size: var(--sl-input-font-size-large);
     min-height: var(--sl-input-height-large);
@@ -271,15 +271,15 @@ export default css`
   /*
    * Pill modifier
    */
-  .select--pill.select--small .select__box {
+  .select--pill.select--small .select__control {
     border-radius: var(--sl-input-height-small);
   }
 
-  .select--pill.select--medium .select__box {
+  .select--pill.select--medium .select__control {
     border-radius: var(--sl-input-height-medium);
   }
 
-  .select--pill.select--large .select__box {
+  .select--pill.select--large .select__control {
     border-radius: var(--sl-input-height-large);
   }
 `;