+++ src/components/select/select.ts
@@ -44,6 +44,7 @@ let id = 0;
  *
  * @csspart base - The component's base wrapper.
  * @csspart clear-button - The input's clear button, exported from <sl-input>.
+ * @csspart control - The container that holds the prefix, label, and suffix.
  * @csspart form-control - The form control that wraps the label, input, and help text.
  * @csspart help-text - The select's help text.
  * @csspart icon - The select's icon.
@@ -59,7 +60,7 @@ export default class SlSelect extends LitElement {
   static styles = styles;
 
   @query('.select') dropdown: SlDropdown;
-  @query('.select__box') box: SlDropdown;
+  @query('.select__control') control: SlDropdown;
   @query('.select__hidden-select') input: HTMLInputElement;
   @query('.select__menu') menu: SlMenu;
 
@@ -266,7 +267,7 @@ export default class SlSelect extends LitElement {
   }
 
   handleLabelClick() {
-    const box = this.shadowRoot?.querySelector('.select__box') as HTMLElement;
+    const box = this.shadowRoot?.querySelector('.select__control') as HTMLElement;
     box.focus();
   }
 
@@ -293,7 +294,7 @@ export default class SlSelect extends LitElement {
     this.isOpen = false;
 
     // Restore focus on the box after the menu is hidden
-    this.box.focus();
+    this.control.focus();
   }
 
   @watch('multiple')
@@ -351,7 +352,7 @@ export default class SlSelect extends LitElement {
   }
 
   resizeMenu() {
-    const box = this.shadowRoot?.querySelector('.select__box') as HTMLElement;
+    const box = this.shadowRoot?.querySelector('.select__control') as HTMLElement;
     this.menu.style.width = `${box.clientWidth}px`;
 
     if (this.dropdown) {
@@ -464,9 +465,10 @@ export default class SlSelect extends LitElement {
           @sl-hide=${this.handleMenuHide}
         >
           <div
+            part="control"
             slot="trigger"
             id=${this.inputId}
-            class="select__box"
+            class="select__control"
             role="combobox"
             aria-labelledby=${ifDefined(
               getLabelledBy({