+++ docs/resources/changelog.md
@@ -10,6 +10,8 @@ _During the beta period, these restrictions may be relaxed in the event of a mis
 
 - Added initial surface design tokens
 - Added missing background color to `<sl-details>`
+- Added the `control` part to `<sl-select>` so you can target the main control with CSS [#538](https://github.com/shoelace-style/shoelace/issues/538)
+- Changed the default `distance` in `<sl-dropdown>` from `2` to `0` [#538](https://github.com/shoelace-style/shoelace/issues/538)
 - Modified the color scale to no longer use a luminance shift
 
 ## 2.0.0-beta.52
@@ -19,7 +21,6 @@ _During the beta period, these restrictions may be relaxed in the event of a mis
 - Added the `--speed` custom property to `<sl-spinner>`
 - Added the `--size` and `--track-width` custom properties to `<sl-progress-ring>`
 - Added tests for `<sl-badge>` [#530](https://github.com/shoelace-style/shoelace/pull/530)
-- Changed the default `distance` in `<sl-dropdown>` from `2` to `0`
 - Fixed a bug where `<sl-tab>` wasn't using a border radius token [#523](https://github.com/shoelace-style/shoelace/issues/523)
 - Fixed a bug in the Remix Icons example where some icons would 404 [#528](https://github.com/shoelace-style/shoelace/issues/528)
 - Updated `<sl-progress-ring>` to use only CSS for styling